TilesAndColor
Add support for tiles json files
Add color to points

+++ Assets/Scripts/Model/IPoint.cs
@@ -3,5 +3,5 @@ public interface IPoint<T>
     public T X { get; set; }
     public T Y { get; set; }
     public T Z { get; set; }
-    public uint[] Color { get; set; }
+    public byte[] Color { get; set; }
 }
